+++ src/SalesManager.java
@@ -15,9 +15,9 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = Integer.MAX_VALUE;
-        for (int sale : sales) {
+    public long min() {
+        long min = Integer.MAX_VALUE;
+        for (long sale : sales) {
             if (min > sale) {
                 min = sale;
             }
@@ -26,9 +26,9 @@ public class SalesManager {
         return min;
     }
 
-    public int average() {
-        int sum = 0;
-        for (int sale : sales) {
+    public long average() {
+        long sum = 0;
+        for (long sale : sales) {
             sum += sale;
         }
